+++ src/view/TowerView.java
@@ -11,7 +11,7 @@ public class TowerView {
     private static HashMap<TowerModel, BufferedImage> towerSprites = new HashMap<TowerModel, BufferedImage>();
 
     public static void renderTower(Graphics g, TowerModel tower){
-        g.drawImage(getSprite(tower), tower.getX()*AppView.UNIT_SIZE, tower.getY()*AppView.UNIT_SIZE, null);
+        g.drawImage(getSprite(tower), (int) tower.getX()*AppView.UNIT_SIZE, (int) tower.getY()*AppView.UNIT_SIZE, null);
     }
 
     public static void renderTowerAt(Graphics g, TowerModel tower, int x, int y){
